remove duplicate fq2_push fq12_push
