+++ src/bn254/fq6.rs
@@ -611,31 +611,14 @@ impl Fq6 {
 
 #[cfg(test)]
 mod test {
-    use crate::bn254::fp254impl::Fp254Impl;
-    use crate::bn254::fq::Fq;
     use crate::bn254::fq6::Fq6;
+    use crate::bn254::utils::{fq2_push, fq6_push};
     use crate::treepp::*;
     use ark_ff::Field;
     use ark_std::UniformRand;
     use core::ops::Mul;
-    use num_bigint::BigUint;
     use rand::SeedableRng;
     use rand_chacha::ChaCha20Rng;
-
-    fn fq2_push(element: ark_bn254::Fq2) -> Script {
-        script! {
-            { Fq::push_u32_le(&BigUint::from(element.c0).to_u32_digits()) }
-            { Fq::push_u32_le(&BigUint::from(element.c1).to_u32_digits()) }
-        }
-    }
-
-    fn fq6_push(element: ark_bn254::Fq6) -> Script {
-        script! {
-            for elem in element.to_base_prime_field_elements() {
-                { Fq::push_u32_le(&BigUint::from(elem).to_u32_digits()) }
-           }
-        }
-    }
 
     #[test]
     fn test_bn254_fq6_add() {
