+++ src/bn254/mod.rs
@@ -12,3 +12,4 @@ pub mod ell_coeffs;
 
 pub mod fp254impl;
 pub mod msm;
+pub mod utils;
